Fixed tests and moved patches around
